# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Poll App API
  description: |
    API for the Poll Application with authentication, poll management, and voting functionality.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    Use the refresh token endpoint to obtain new access tokens when they expire.
  version: 1.0.0
  contact:
    name: Poll App Team

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8080/api
    description: Local development server (with /api prefix)

tags:
  - name: users
    description: User management and authentication
  - name: polls
    description: Poll management
  - name: votes
    description: Voting functionality
  - name: health
    description: Health check

paths:
  /health:
    get:
      tags:
        - health
      summary: Health check
      description: Returns the health status of the API
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
                example: "We are Up!"

  /api/users:
    post:
      tags:
        - users
      summary: Create a new user
      description: Register a new user account
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/users/login:
    post:
      tags:
        - users
      summary: User login
      description: Authenticate user and receive access tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/users/refresh:
    post:
      tags:
        - users
      summary: Refresh access token
      description: Get a new access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/polls:
    get:
      tags:
        - polls
      summary: List all polls
      description: Get a list of all polls
      operationId: listPolls
      responses:
        '200':
          description: List of polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PollResponse'
    post:
      tags:
        - polls
      summary: Create a new poll
      description: Create a new poll (requires authentication)
      operationId: createPoll
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePollRequest'
      responses:
        '201':
          description: Poll created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/polls/{id}:
    get:
      tags:
        - polls
      summary: Get poll by ID
      description: Get detailed information about a specific poll
      operationId: getPoll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Poll ID
      responses:
        '200':
          description: Poll details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollResponse'
        '404':
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - polls
      summary: Update poll
      description: Update a poll (requires authentication and ownership)
      operationId: updatePoll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Poll ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePollRequest'
      responses:
        '200':
          description: Poll updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollResponse'
        '400':
          description: Invalid request or not owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - polls
      summary: Delete poll
      description: Delete a poll (requires authentication and ownership)
      operationId: deletePoll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Poll ID
      responses:
        '204':
          description: Poll deleted successfully
        '400':
          description: Invalid request or not owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/polls/{id}/vote:
    post:
      tags:
        - votes
      summary: Vote on a poll
      description: Submit a vote for a poll option (requires authentication, one vote per user per poll)
      operationId: voteOnPoll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Poll ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      responses:
        '201':
          description: Vote submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '400':
          description: Invalid request or already voted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/polls/{id}/votes:
    get:
      tags:
        - votes
      summary: Get vote counts
      description: Get vote counts for all options in a poll
      operationId: getVoteCounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Poll ID
      responses:
        '200':
          description: Vote counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteCountsResponse'
        '404':
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/polls/{id}/votes/{option}:
    get:
      tags:
        - votes
      summary: Get voters by option
      description: Get list of users who voted for a specific option
      operationId: getVotersByOption
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Poll ID
        - name: option
          in: path
          required: true
          schema:
            type: string
          description: Poll option
      responses:
        '200':
          description: List of voters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VotersResponse'
        '404':
          description: Poll or option not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

  schemas:
    CreateUserRequest:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          minLength: 1
          example: johndoe
        password:
          type: string
          format: password
          minLength: 1
          example: securepassword123

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: securepassword123

    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (15 minutes TTL)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token (7 days TTL)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe

    CreatePollRequest:
      type: object
      required:
        - title
        - options
      properties:
        title:
          type: string
          minLength: 1
          example: What's your favorite programming language?
        description:
          type: string
          example: Please select your preferred programming language
        options:
          type: array
          minItems: 2
          items:
            type: string
          example:
            - Go
            - JavaScript
            - Python
            - Rust

    UpdatePollRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: What's your favorite programming language?
        description:
          type: string
          example: Please select your preferred programming language
        options:
          type: array
          minItems: 2
          items:
            type: string
          example:
            - Go
            - JavaScript
            - Python
            - Rust

    PollResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          example: What's your favorite programming language?
        description:
          type: string
          nullable: true
          example: Please select your preferred programming language
        options:
          type: array
          items:
            type: string
          example:
            - Go
            - JavaScript
            - Python
            - Rust
        owner_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        vote_counts:
          type: object
          additionalProperties:
            type: integer
          description: Map of option to vote count (only included in poll details endpoint)
          example:
            Go: 10
            JavaScript: 25
            Python: 15
            Rust: 5
        voters_by_option:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/UserInfo'
          description: Map of option to list of voters (only included in poll details endpoint)
          example:
            Go:
              - id: 123e4567-e89b-12d3-a456-426614174000
                email: user1@example.com
                username: user1
              - id: 223e4567-e89b-12d3-a456-426614174001
                email: user2@example.com
                username: user2
            JavaScript:
              - id: 323e4567-e89b-12d3-a456-426614174002
                email: user3@example.com
                username: user3

    VoteRequest:
      type: object
      required:
        - option
      properties:
        option:
          type: string
          minLength: 1
          example: Go

    VoteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        user_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        poll_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        option:
          type: string
          example: Go
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"

    VoteCountsResponse:
      type: object
      properties:
        poll_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        counts:
          type: object
          additionalProperties:
            type: integer
          description: Map of option to vote count
          example:
            Go: 10
            JavaScript: 25
            Python: 15
            Rust: 5

    VotersResponse:
      type: object
      properties:
        poll_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        option:
          type: string
          example: Go
        voters:
          type: array
          items:
            $ref: '#/components/schemas/UserInfo'
          example:
            - id: 123e4567-e89b-12d3-a456-426614174000
              email: user1@example.com
              username: user1
            - id: 223e4567-e89b-12d3-a456-426614174001
              email: user2@example.com
              username: user2

    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Invalid request